+++ src/config.js
@@ -1,4 +1,4 @@
-const constants = require("./constants");
+const { HTTP_CLIENT, TS_KEYWORDS, PRETTIER_OPTIONS } = require("./constants");
 
 const config = {
   /** CLI flag */
@@ -37,7 +37,7 @@ const config = {
     outOfModuleApi: "Common",
   },
   routeNameDuplicatesMap: new Map(),
-  prettierOptions: constants.PRETTIER_OPTIONS,
+  prettierOptions: PRETTIER_OPTIONS,
   hooks: {
     onCreateComponent: (schema) => schema,
     onParseSchema: (originalSchema, parsedSchema) => parsedSchema,
@@ -47,8 +47,23 @@ const config = {
     onCreateRequestParams: (rawType) => {},
     onCreateRouteName: () => {},
   },
-  defaultResponseType: constants.TS_KEYWORDS.VOID,
+  defaultResponseType: TS_KEYWORDS.VOID,
   singleHttpClient: false,
+  httpClientType: HTTP_CLIENT.FETCH,
+  templatePaths: {
+    /** `templates/base` */
+    base: "",
+    /** `templates/default` */
+    default: "",
+    /** `templates/modular` */
+    modular: "",
+    /** usage path if `--templates` option is not set */
+    original: "",
+    /** custom path to templates (`--templates`) */
+    custom: "",
+  },
+  /** Record<templateName, templateContent> */
+  templatesToRender: {},
 };
 
 /** needs to use data everywhere in project */